+++ src/components/select/index.js
@@ -1,3 +1,7 @@
 import Select from './Select'
+import Option from './Option'
 
-export default Select
+export {
+    Select,
+    Option
+}